Move memory and table styles into wasmer_types
