+++ lib/api/src/sys/tunables.rs
@@ -8,9 +8,10 @@ use wasmer_compiler::Target;
 use wasmer_engine::Tunables;
 use wasmer_vm::MemoryError;
 use wasmer_vm::{
-    LinearMemory, LinearTable, Memory, MemoryStyle, Table, TableStyle, VMMemoryDefinition,
+    LinearMemory, LinearTable, Memory, Table, VMMemoryDefinition,
     VMTableDefinition,
 };
+use wasmer_types::{MemoryStyle, TableStyle};
 
 /// Tunable parameters for WebAssembly compilation.
 /// This is the reference implementation of the `Tunables` trait,
