+++ lib/engine-staticlib/src/artifact.rs
@@ -25,10 +25,10 @@ use wasmer_types::entity::{BoxedSlice, PrimaryMap};
 use wasmer_types::DataInitializer;
 use wasmer_types::{
     FunctionIndex, LocalFunctionIndex, MemoryIndex, ModuleInfo, OwnedDataInitializer,
-    SignatureIndex, TableIndex,
+    SignatureIndex, TableIndex, MemoryStyle, TableStyle, 
 };
 use wasmer_vm::{
-    FuncDataRegistry, FunctionBodyPtr, MemoryStyle, TableStyle, VMSharedSignatureIndex,
+    FuncDataRegistry, FunctionBodyPtr, VMSharedSignatureIndex,
     VMTrampoline,
 };
