+++ lib/engine-universal/src/artifact.rs
@@ -20,10 +20,10 @@ use wasmer_engine::{Engine, Tunables};
 use wasmer_types::entity::{BoxedSlice, PrimaryMap};
 use wasmer_types::{
     FunctionIndex, LocalFunctionIndex, MemoryIndex, ModuleInfo, OwnedDataInitializer,
-    SignatureIndex, TableIndex,
+    SignatureIndex, TableIndex, MemoryStyle, TableStyle, 
 };
 use wasmer_vm::{
-    FuncDataRegistry, FunctionBodyPtr, MemoryStyle, TableStyle, VMSharedSignatureIndex,
+    FuncDataRegistry, FunctionBodyPtr, VMSharedSignatureIndex,
     VMTrampoline,
 };
