+++ lib/engine/src/tunables.rs
@@ -5,11 +5,10 @@ use std::sync::Arc;
 use wasmer_types::entity::{EntityRef, PrimaryMap};
 use wasmer_types::{
     GlobalType, LocalGlobalIndex, LocalMemoryIndex, LocalTableIndex, MemoryIndex, MemoryType,
-    ModuleInfo, TableIndex, TableType,
+    ModuleInfo, TableIndex, TableType, MemoryStyle, TableStyle
 };
 use wasmer_vm::MemoryError;
 use wasmer_vm::{Global, Memory, Table};
-use wasmer_vm::{MemoryStyle, TableStyle};
 use wasmer_vm::{VMMemoryDefinition, VMTableDefinition};
 
 /// An engine delegates the creation of memories, tables, and globals
