+++ lib/types/src/lib.rs
@@ -63,6 +63,7 @@ mod initializers;
 mod memory_view;
 mod module;
 mod native;
+mod styles;
 mod trapcode;
 mod types;
 mod units;
@@ -88,6 +89,7 @@ pub use crate::units::{
     Bytes, PageCountOutOfRange, Pages, WASM_MAX_PAGES, WASM_MIN_PAGES, WASM_PAGE_SIZE,
 };
 pub use crate::values::{Value, WasmValueType};
+pub use crate::styles::{MemoryStyle, TableStyle};
 pub use types::{
     ExportType, ExternType, FunctionType, GlobalInit, GlobalType, ImportType, MemoryType,
     Mutability, TableType, Type, V128,
